requirements.txt
pyracing==0.8.2
Pillow==8.3.2
httpx==0.19.0